+++ RELEASE-NOTES.md
@@ -24,7 +24,8 @@ Picocli follows [semantic versioning](http://semver.org/).
 * [#644][#671] Auto-completion: fix [shellcheck](https://github.com/koalaman/shellcheck) warnings in generated autocompletion scripts. Thanks to [Dylan Cali](https://github.com/calid) for raising this, and thanks to [AlcaYezz](https://github.com/AlcaYezz) for the pull request.
 * [#396] Auto-completion: completion scripts no longer use associative arrays, and should now work on OSX.
 * [#629] Support injecting `@Spec CommandSpec spec` into `IVersionProvider` implementations. Thanks to [Garret Wilson](https://github.com/garretwilson) for raising this.
-* [#926] Clarify debug trace output when adding aliases
+* [#926] Enhancement: Clarify debug trace output when adding aliases.
+* [#928] Enhancement: Improve debug tracing: show command user object identity hashcode and prefix "Processing argument..." with argument index.
 * [#925] Enhancement: Support `@ParentCommand`-annotated fields in mixin classes.
 * [#690] Enhancement: Postpone instantiating subcommands until they are matched on the command line. Thanks to [Daniel Breitlauch](https://github.com/danielBreitlauch) for raising this.
 * [#920] Enhancement: Reduce `DEBUG` tracing noise if no resource bundle is set.